backend/quick_test.py
#!/usr/bin/env python3
"""
Quick test to verify your legal voice app is working
"""

import requests
import json
import time

def test_app():
    print("🧪 Testing Legal Voice Application...")
    print("=" * 50)
    
    # Test if backend is running
    try:
        response = requests.get("http://localhost:8000/", timeout=5)
        print("✅ Backend is running on http://localhost:8000")
    except:
        print("❌ Backend not running. Please start with: python start_app.py")
        return
    
    # Test if frontend is running
    try:
        response = requests.get("http://localhost:3000/", timeout=5)
        print("✅ Frontend is running on http://localhost:3000")
    except:
        print("⚠️ Frontend not running. Please start with: npm run dev")
    
    # Test OpenAI integration
    print("\n🔍 Testing OpenAI Integration...")
    
    # Test form interpretation
    test_data = {
        "form_id": "name_change",
        "transcript": "My name is John Doe, I am 30 years old, my father is Robert Doe"
    }
    
    try:
        # Note: This will fail without authentication, but shows the endpoint exists
        response = requests.post("http://localhost:8000/interpret", 
                               json=test_data, timeout=10)
        if response.status_code == 401:
            print("✅ OpenAI endpoints are available (authentication required)")
        else:
            print(f"✅ OpenAI integration working: {response.status_code}")
    except Exception as e:
        print(f"⚠️ OpenAI test: {str(e)}")
    
    print("\n🎯 How to Test Manually:")
    print("1. Open http://localhost:3000 in your browser")
    print("2. Select a form (e.g., Name Change Affidavit)")
    print("3. Click the microphone and speak")
    print("4. Watch the AI fill the form automatically!")
    
    print("\n📱 Test Scenarios:")
    print("• English: 'My name is John Doe, I am 30 years old'")
    print("• Hindi: 'मेरा नाम राम शर्मा है, मैं 25 साल का हूं'")
    print("• Mixed: 'My name is Rajesh, मैं 28 साल का हूं'")
    
    print("\n🎉 Your Legal Voice App is Ready!")

if __name__ == "__main__":
    test_app()
