backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pymongo==4.6.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
python-dotenv==1.0.0
pyjwt==2.8.1
openai==1.3.0
sendgrid==6.10.0
twilio==8.10.0
python-multipart==0.0.6
aiofiles==23.2.1
websockets==12.0
reportlab==4.0.7
cloudinary==1.36.0
